ensuring tests pass. Also updated the setup.rb to be more robust.
NOTE: there is still one failing test

test_weights(SearchTest) [./test/integration/search_test.rb:310]:
<#<Seller id: 16, user_id: 16, company_name: "seller16", created_at: "2007-07-03 04:28:11", updated_at: "2007-10-07 04:28:11", capitalization: 24.768, mission_statement: "superior developing through value Add products.">> expected to be != to
<#<Seller id: 16, user_id: 16, company_name: "seller16", created_at: "2007-07-03 04:28:11", updated_at: "2007-10-07 04:28:11", capitalization: 24.768, mission_statement: "superior developing through value Add products.">>.

However, there I have no explanation for why this would fail due to the escaping.

+++ test/setup.rb
@@ -6,7 +6,9 @@ Dir.chdir "#{File.dirname(__FILE__)}/integration/app/" do
   pid_file = '/tmp/sphinx/searchd.pid'
   if File.exist? pid_file
     pid = File.read(pid_file).to_i
-    system("kill #{pid}"); sleep(2); system("kill -9 #{pid}")  
+    if pid > 0
+      system("kill #{pid}"); sleep(2); system("kill -9 #{pid}")  
+    end
   end
   
   system("rm -rf /tmp/sphinx")
